+++ src/assembly/getQueenHitMoves.ts
@@ -1,5 +1,5 @@
-export function getQueenHitMoves(position: i8, board: Uint8Array, color: u8): u16[] {
-  const targets: u16[] = [];
+export function getQueenHitMoves(position: i8, board: Uint8Array, color: u8): u8[] {
+  const targets: u8[] = [];
 
   const maxMovesUp = position >>> 3; // equals to rank
   const maxMovesLeft = position & 7; // equals to file
